# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Assuredia QA Monitoring Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ClientStatus = (typeof ClientStatus)[keyof typeof ClientStatus];

export const ClientStatus = {
  active: "active",
  inactive: "inactive",
  warning: "warning",
  error: "error",
} as const;

export interface Client {
  id: number;
  name: string;
  environment: string;
  status: ClientStatus;
  successRate: number;
  lastRunAt?: string | null;
  webhookUrl?: string | null;
  createdAt: string;
}

export interface CreateClientBody {
  name: string;
  environment: string;
  webhookUrl?: string | null;
}

export type UpdateClientBodyStatus =
  (typeof UpdateClientBodyStatus)[keyof typeof UpdateClientBodyStatus];

export const UpdateClientBodyStatus = {
  active: "active",
  inactive: "inactive",
  warning: "warning",
  error: "error",
} as const;

export interface UpdateClientBody {
  name?: string;
  environment?: string;
  status?: UpdateClientBodyStatus;
  webhookUrl?: string | null;
}

export type TestRunStatus = (typeof TestRunStatus)[keyof typeof TestRunStatus];

export const TestRunStatus = {
  running: "running",
  passed: "passed",
  failed: "failed",
  skipped: "skipped",
} as const;

export interface TestRun {
  id: number;
  clientId: number;
  clientName: string;
  flow: string;
  status: TestRunStatus;
  totalTests: number;
  passed: number;
  failed: number;
  skipped: number;
  duration?: number | null;
  startedAt: string;
  completedAt?: string | null;
}

export interface CreateTestRunBody {
  clientId: number;
  flow: string;
  webhookUrl?: string | null;
}

export type AlertSeverity = (typeof AlertSeverity)[keyof typeof AlertSeverity];

export const AlertSeverity = {
  critical: "critical",
  warning: "warning",
  info: "info",
} as const;

export interface Alert {
  id: number;
  clientId: number;
  clientName: string;
  testRunId?: number | null;
  severity: AlertSeverity;
  message: string;
  module?: string | null;
  resolved: boolean;
  createdAt: string;
}

export type CreateAlertBodySeverity =
  (typeof CreateAlertBodySeverity)[keyof typeof CreateAlertBodySeverity];

export const CreateAlertBodySeverity = {
  critical: "critical",
  warning: "warning",
  info: "info",
} as const;

export interface CreateAlertBody {
  clientId: number;
  testRunId?: number | null;
  severity: CreateAlertBodySeverity;
  message: string;
  module?: string | null;
}

export interface DashboardSummary {
  successRate: number;
  totalTests: number;
  failedTests: number;
  activeRuns: number;
  totalClients: number;
  unresolvedAlerts: number;
}

export interface TrendPoint {
  date: string;
  passed: number;
  failed: number;
  skipped: number;
}

export interface ModuleFailure {
  module: string;
  count: number;
}

export interface Report {
  id: number;
  clientId: number;
  clientName: string;
  period: string;
  totalRuns: number;
  avgSuccessRate: number;
  createdAt: string;
}

export type ListTestRunsParams = {
  clientId?: number;
  limit?: number;
};

export type ListAlertsParams = {
  clientId?: number;
  limit?: number;
};

export type GetExecutionTrendParams = {
  days?: number;
};

export type ListReportsParams = {
  clientId?: number;
};