# lib/api-zod/src/generated/types/alert.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Assuredia QA Monitoring Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AlertSeverity } from "./alertSeverity";

export interface Alert {
  id: number;
  clientId: number;
  clientName: string;
  testRunId?: number | null;
  severity: AlertSeverity;
  message: string;
  module?: string | null;
  resolved: boolean;
  createdAt: string;
}