# lib/api-zod/src/generated/types/createTestRunBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Assuredia QA Monitoring Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateTestRunBody {
  clientId: number;
  flow: string;
  webhookUrl?: string | null;
}