# lib/api-zod/src/generated/types/updateClientBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Assuredia QA Monitoring Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateClientBodyStatus } from "./updateClientBodyStatus";

export interface UpdateClientBody {
  name?: string;
  environment?: string;
  status?: UpdateClientBodyStatus;
  webhookUrl?: string | null;
}